requirements.txt
allure-pytest==2.15.0
pytest==8.4.2
requests==2.32.5
selenium==4.36.0
webdriver-manager==4.0.2

#faker
#flake8